+++ cpskin/core/viewlets/banner.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 import random
 
+from Acquisition import aq_inner
 from plone import api
+from plone.app.layout.navigation.interfaces import INavigationRoot
 from plone.app.layout.viewlets.common import ViewletBase
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from zope.component import getMultiAdapter
@@ -30,6 +32,27 @@ class CPSkinBannerViewlet(ViewletBase):
             return False
         return (self.isInMinisiteMode() or self.isInPortalMode())
 
+    def show_slogan(self):
+        portal_registry = api.portal.get_tool('portal_registry')
+        show_slogan = portal_registry[
+            'cpskin.core.interfaces.ICPSkinSettings.show_slogan']
+        return show_slogan
+
+    def slogan(self):
+        default = {
+            'title': "",
+            'description': "",
+        }
+        if not self.isHomepage():
+            return default
+        banner = self.getBanner()
+        if not banner:
+            return default
+        return {
+            'title': banner.Title(),
+            'description': banner.Description(),
+        }
+
     def homeUrl(self):
         """
         Returns URL of :
@@ -48,6 +71,12 @@ class CPSkinBannerViewlet(ViewletBase):
             minisiteRoot = portal.unrestrictedTraverse(minisite.search_path)
             return minisiteRoot.absolute_url()
 
+    def isHomepage(self):
+        if self.isInMinisite():
+            return False
+        obj = aq_inner(self.context)
+        return INavigationRoot.providedBy(obj)
+
     def isInMinisiteMode(self):
         if not HAS_MINISITE:
             return False
@@ -60,16 +89,20 @@ class CPSkinBannerViewlet(ViewletBase):
         request = self.request
         return IInPortal.providedBy(request)
 
-    def getImageBannerUrl(self):
+    def getBanner(self):
         context = self.context
         banner_folder = getattr(context, 'banner', None)
-        image = ''.join([context.absolute_url(), '/banner.jpg'])
+        banner = getattr(context, 'banner.jpg', None)
         if banner_folder:
             images = api.content.find(
                 context=banner_folder,
                 portal_type='Image',
             )
             if images:
-                item = images[random.randrange(len(images))]
-                image = item.getURL()
-        return image
+                banner = images[random.randrange(len(images))]
+                return banner.getObject()
+        return banner
+
+    def getImageBannerUrl(self):
+        banner = self.getBanner()
+        return banner and banner.absolute_url() or ''